requirements.txt
# cat requirements.txt
biopython   
scipy  
numpy
pomegranate==0.14.9  
